+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Fixed
 - `_smoothen_cell` returns empty polygon if the cell can't be smoothen (#279)
+- Broken link in the docs @ChristopherBottomsOMRF (#284)
 
 ## [2.1.0] - 2025-06-27