+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Fixed
 - `_smoothen_cell` returns empty polygon if the cell can't be smoothen (#279)
 - Broken link in the docs @ChristopherBottomsOMRF (#284)
+- Added again command `ps` to all Docker images for Nextflow
 
 ## [2.1.0] - 2025-06-27