+++ CHANGELOG.md
@@ -1,9 +1,9 @@
-## [2.1.1] - xxxx-xx-xx
+## [2.1.1] - 2025-07-16
 
 ### Added
 - Log a warning in `sopa.io.visium_hd` if the fullres image is too small (potentially a user error)
 - Added a `allow_holes` argument to `sopa.segmentation.tissue` to decide whether to keep holes or not
-- `correction` argument in `sopa.spatial.mean_distance` to account for the bias related to groups proportions
+- `correction` argument in `sopa.spatial.mean_distance` to account for the bias related to groups proportions (experimental)
 - The Docker CI now also pushes the images with the `latest` tag
 
 ### Changed